Add support for test names that contain data
In some cases test names include data in them. For e.g. when using xUnit you can create the so-called 'Theory' tests which run a given test with different inputs generated by a static method. In this case the test platform returns individual tests with different inputs. This commit fixes the parsing the logic to allow for such tests to be included in the test explorer.