Detect test results files from dotnet watch
Removed the class "TestResultsFile" in the process, since it contained
only a single method and no state; made the method a free function
instead.

Also fixed a bug related to the parsing of the `dotnet watch` process
output - since the callback may not return whole lines, we have to
split and join them into lines by ourselves.

The method `TestCommands.watchRunningTests` was also removed to avoid
confusion about it; it did not actually watch any running tests.